+++ telnetlib3/tests/test_linemode.py
@@ -53,7 +53,12 @@ async def test_server_demands_remote_linemode_client_agrees(  # pylint: disable=
             client_writer.write(reply_stage2)
 
             srv_instance = await asyncio.wait_for(server.wait_for_client(), 0.1)
-            assert not any(srv_instance.writer.pending_option.values())
+            await asyncio.wait_for(
+                srv_instance.writer.wait_for(
+                    remote={"LINEMODE": True}, pending={"LINEMODE": False}
+                ),
+                0.1,
+            )
 
             result = await client_reader.read()
             assert result == b""
@@ -65,7 +70,6 @@ async def test_server_demands_remote_linemode_client_agrees(  # pylint: disable=
             assert srv_instance.writer.linemode.ack is True
             assert srv_instance.writer.linemode.soft_tab is False
             assert srv_instance.writer.linemode.lit_echo is True
-            assert srv_instance.writer.remote_option.enabled(LINEMODE)
 
 
 async def test_server_demands_remote_linemode_client_demands_local(  # pylint: disable=too-many-locals
@@ -109,7 +113,12 @@ async def test_server_demands_remote_linemode_client_demands_local(  # pylint: d
             client_writer.write(reply_stage2)
 
             srv_instance = await asyncio.wait_for(server.wait_for_client(), 0.1)
-            assert not any(srv_instance.writer.pending_option.values())
+            await asyncio.wait_for(
+                srv_instance.writer.wait_for(
+                    remote={"LINEMODE": True}, pending={"LINEMODE": False}
+                ),
+                0.1,
+            )
 
             result = await client_reader.read()
             assert result == b""
@@ -121,4 +130,3 @@ async def test_server_demands_remote_linemode_client_demands_local(  # pylint: d
             assert srv_instance.writer.linemode.ack is True
             assert srv_instance.writer.linemode.soft_tab is False
             assert srv_instance.writer.linemode.lit_echo is False
-            assert srv_instance.writer.remote_option.enabled(LINEMODE)